LPS-179290 Extract on click
