+++ modules/apps/site-navigation/site-navigation-admin-web/src/main/resources/META-INF/resources/site_navigation_menu_editor/components/AddItemDropdown.js
@@ -13,193 +13,40 @@
  */
 
 import ClayDropDown from '@clayui/drop-down';
-import {
-	fetch,
-	objectToFormData,
-	openModal,
-	openSelectionModal,
-} from 'frontend-js-web';
 import PropTypes from 'prop-types';
 import React, {useState} from 'react';
 
 import {useConstants} from '../contexts/ConstantsContext';
 
-function getNamespacedInfoItem(
-	portletNamespace,
-	selectedItem,
-	siteNavigationMenuId
-) {
-	if (!selectedItem) {
-		return;
-	}
-
-	let infoItem = {
-		...selectedItem,
-	};
-
-	let value;
-
-	if (typeof selectedItem.value === 'string') {
-		try {
-			value = JSON.parse(selectedItem.value);
-		}
-		catch (error) {}
-	}
-	else if (selectedItem.value && typeof selectedItem.value === 'object') {
-		value = selectedItem.value;
-	}
-
-	if (value) {
-		delete infoItem.value;
-		infoItem = {...value};
-	}
-
-	infoItem.siteNavigationMenuId = siteNavigationMenuId;
-
-	return Liferay.Util.ns(portletNamespace, infoItem);
-}
-
-function getNamespacedInfoItems(
-	portletNamespace,
-	selectedItems,
-	siteNavigationMenuId
-) {
-	if (!selectedItems) {
-		return;
-	}
-
-	let selectedItemsValue = selectedItems;
-
-	if (selectedItems.value && Array.isArray(selectedItems.value)) {
-		selectedItemsValue = selectedItems.value.map((item) =>
-			JSON.parse(item)
-		);
-	}
-	else if (typeof selectedItems === 'object') {
-		selectedItemsValue = Object.values(selectedItems);
-	}
-
-	if (!selectedItemsValue.length) {
-		return;
-	}
-
-	const infoItems = {
-		items: JSON.stringify(selectedItemsValue),
-		siteNavigationMenuId,
-	};
-
-	return Liferay.Util.ns(portletNamespace, infoItems);
-}
-
 export function AddItemDropDown({className, trigger}) {
 	const [active, setActive] = useState(false);
-	const {addSiteNavigationMenuItemOptions, portletNamespace} = useConstants();
-
-	const onItemAdd = () => {
-		window.sessionStorage.setItem(`${portletNamespace}itemAdded`, true);
-	};
+	const {addSiteNavigationMenuItemOptions} = useConstants();
 
 	return (
 		<>
 			<ClayDropDown
 				active={active}
 				className={className}
+				menuElementAttrs={{
+					containerProps: {
+						className: 'add-navigation-menu-item-dropdown',
+					},
+				}}
 				onActiveChange={setActive}
 				trigger={trigger}
 			>
 				<ClayDropDown.ItemList>
-					{addSiteNavigationMenuItemOptions.map(({data, label}) => (
-						<ClayDropDown.Item
-							key={label}
-							onClick={() => {
-								const useSmallerModal = shouldUseSmallerModal(
-									data.type
-								);
-
-								if (data.itemSelector) {
-									openSelectionModal({
-										buttonAddLabel: data.multiSelection
-											? Liferay.Language.get('select')
-											: null,
-										height: useSmallerModal
-											? '60vh'
-											: undefined,
-										multiple: data.multiSelection,
-
-										onSelect: (selection) => {
-											fetch(data.addItemURL, {
-												body: objectToFormData(
-													data.multiSelection
-														? getNamespacedInfoItems(
-																portletNamespace,
-																selection,
-																data.siteNavigationMenuId
-														  )
-														: getNamespacedInfoItem(
-																portletNamespace,
-																selection,
-																data.siteNavigationMenuId
-														  )
-												),
-												method: 'POST',
-											}).then(() => {
-												onItemAdd();
-
-												window.location.reload();
-											});
-										},
-
-										selectEventName: `${portletNamespace}selectItem`,
-										size: useSmallerModal
-											? 'md'
-											: undefined,
-										title: data.addTitle,
-										url: data.href,
-									});
-								}
-								else {
-									openModal({
-										height: useSmallerModal
-											? '60vh'
-											: undefined,
-										id: `${portletNamespace}addMenuItem`,
-										iframeBodyCssClass: 'portal-popup',
-										size: useSmallerModal
-											? 'md'
-											: undefined,
-										title: data.addTitle,
-										url: data.href,
-									});
-
-									Liferay.once(
-										'reloadSiteNavigationMenuEditor',
-										() => {
-											onItemAdd();
-
-											window.location.reload();
-										}
-									);
-								}
-							}}
-						>
-							{label}
-						</ClayDropDown.Item>
-					))}
+					{addSiteNavigationMenuItemOptions.map(
+						({label, onClick}) => (
+							<ClayDropDown.Item key={label} onClick={onClick}>
+								{label}
+							</ClayDropDown.Item>
+						)
+					)}
 				</ClayDropDown.ItemList>
 			</ClayDropDown>
 		</>
 	);
-}
-
-const SMALLER_MODAL_TYPES = [
-	'com.liferay.asset.kernel.model.AssetCategory',
-	'layout',
-	'node',
-	'url',
-];
-
-function shouldUseSmallerModal(type) {
-	return SMALLER_MODAL_TYPES.includes(type);
 }
 
 AddItemDropDown.propTypes = {
